feature/actu

+++ src/main/java/com/mf/mf/model/excel/MFEstadoSituacionFinanciera.java
@@ -14,87 +14,87 @@ public class MFEstadoSituacionFinanciera {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private Boolean estado;
-    private Integer totalActivosCorrientes;
-    private Integer efectivoYEquivalentesAlEfectivo;
-    private Integer efectivoRestringido;
-    private Integer inversionesCortoPlazo;
-    private Integer cuentasComercialesCobrarOperacionalesClientes;
-    private Integer cuentasPorCobrarConPartesRelacionadas1;
-    private Integer activosBiologicos;
-    private Integer otrasCuentasPorCobrar1;
-    private Integer pagosAnticipados;
-    private Integer inventariosCorrientes;
-    private Integer activosPorImpuestos;
-    private Integer activosDistintosAlEfectivoPignoradosComoGarantia;
-    private Integer otrosActivosFinancieros1;
-    private Integer otrosActivosNoFinancieros1;
+    private Long totalActivosCorrientes;
+    private Long efectivoYEquivalentesAlEfectivo;
+    private Long efectivoRestringido;
+    private Long inversionesCortoPlazo;
+    private Long cuentasComercialesCobrarOperacionalesClientes;
+    private Long cuentasPorCobrarConPartesRelacionadas1;
+    private Long activosBiologicos;
+    private Long otrasCuentasPorCobrar1;
+    private Long pagosAnticipados;
+    private Long inventariosCorrientes;
+    private Long activosPorImpuestos;
+    private Long activosDistintosAlEfectivoPignoradosComoGarantia;
+    private Long otrosActivosFinancieros1;
+    private Long otrosActivosNoFinancieros1;
     private String validacionActivosCorrientes;
 
-    private Integer totalActivosNoCorrientes;
-    private Integer depositosYOtrosActivos1;
-    private Integer inversionesLargoPlazo;
-    private Integer cuentasComercialesPorCobrar;
-    private Integer cuentasPorCobrarConPartesRelacionadas2;
-    private Integer otrasCuentasPorCobrar2;
-    private Integer propiedadDeInversion;
-    private Integer activosIntangiblesYCreditoMercantil;
-    private Integer propiedadesPlantaYEquipo;
-    private Integer activosBiologicosNoCorrientes;
-    private Integer inversionesContabilizadasParticipacion;
-    private Integer inversionesSubsidiarNegocios;
-    private Integer plusvalia;
-    private Integer inventariosNoCorrientes;
-    private Integer activosPorImpuestosDiferidos;
-    private Integer activosPorImpuestosCorrientesNoCorrientes;
-    private Integer otrosActivosFinancieros2;
-    private Integer otrosActivosNoFinancieros2;
-    private Integer activosDistintosAlEfectivo;
+    private Long totalActivosNoCorrientes;
+    private Long depositosYOtrosActivos1;
+    private Long inversionesLargoPlazo;
+    private Long cuentasComercialesPorCobrar;
+    private Long cuentasPorCobrarConPartesRelacionadas2;
+    private Long otrasCuentasPorCobrar2;
+    private Long propiedadDeInversion;
+    private Long activosIntangiblesYCreditoMercantil;
+    private Long propiedadesPlantaYEquipo;
+    private Long activosBiologicosNoCorrientes;
+    private Long inversionesContabilizadasParticipacion;
+    private Long inversionesSubsidiarNegocios;
+    private Long plusvalia;
+    private Long inventariosNoCorrientes;
+    private Long activosPorImpuestosDiferidos;
+    private Long activosPorImpuestosCorrientesNoCorrientes;
+    private Long otrosActivosFinancieros2;
+    private Long otrosActivosNoFinancieros2;
+    private Long activosDistintosAlEfectivo;
     private String validacionActivosNoCorrientes; // Cambiado a String (tipo texto)
     private String validacionActivos; // Cambiado a String (tipo texto)
 
-    private Integer pasivos;
-    private Integer pasivosCorrientes;
-    private Integer cuentasPorPagarComercialesProveedores;
-    private Integer otrasCuentasPorPagar;
-    private Integer pasivosPorImpuestos1;
-    private Integer deudaFinanciera;
-    private Integer otrosPasivosNoFinancieros1;
-    private Integer depositosYOtrosActivos2;
-    private Integer porcionCorrienteDeuda;
-    private Integer cuentasPorPagarConPartesRelacionadas3;
-    private Integer otrasProvisiones;
-    private Integer provisionesPorBeneficiosAEmpleados1;
-    private Integer ingresosRecibidosPorCuentaDeTerceros;
+    private Long pasivos;
+    private Long pasivosCorrientes;
+    private Long cuentasPorPagarComercialesProveedores;
+    private Long otrasCuentasPorPagar;
+    private Long pasivosPorImpuestos1;
+    private Long deudaFinanciera;
+    private Long otrosPasivosNoFinancieros1;
+    private Long depositosYOtrosActivos2;
+    private Long porcionCorrienteDeuda;
+    private Long cuentasPorPagarConPartesRelacionadas3;
+    private Long otrasProvisiones;
+    private Long provisionesPorBeneficiosAEmpleados1;
+    private Long ingresosRecibidosPorCuentaDeTerceros;
     private String validacionPasivosCorrientes; // Cambiado a String (tipo texto)
 
-    private Integer pasivosNoCorrientes;
-    private Integer deudaALargoPlazo;
-    private Integer cuentasPorPagar;
-    private Integer provisionesPorBeneficiosAEmpleados2;
-    private Integer pasivosPorImpuestosDiferidos;
-    private Integer ingresosRecibidosPorAnticipado;
-    private Integer otrosPasivosNoFinancieros2;
-    private Integer otrosPasivosFinancieros;
-    private Integer pasivosPorImpuestos2;
+    private Long pasivosNoCorrientes;
+    private Long deudaALargoPlazo;
+    private Long cuentasPorPagar;
+    private Long provisionesPorBeneficiosAEmpleados2;
+    private Long pasivosPorImpuestosDiferidos;
+    private Long ingresosRecibidosPorAnticipado;
+    private Long otrosPasivosNoFinancieros2;
+    private Long otrosPasivosFinancieros;
+    private Long pasivosPorImpuestos2;
     private String validacionPasivosNoCorrientes;
     private String validacionPasivos; // Añadido
 
-    private Integer patrimonio; // Añadido
-    private Integer capitalPagado; // Añadido
-    private Integer primaDeEmision; // Añadido
-    private Integer readquisicionDeInstrumentosDePatrimonioPropio; // Añadido
-    private Integer inversionSumplementariaAlCapitalAsignado; // Añadido
-    private Integer otrasParticipacionesEnElPatrimonio; // Añadido
-    private Integer superavitPorRevaluacion; // Añadido
-    private Integer reservaLegal; // Añadido
-    private Integer otrasReservas; // Añadido
-    private Integer utilidadYorPerdidaDelEjercicio; // Añadido
-    private Integer utilidadYorPerdidaAcumulada; // Añadido
-    private Integer gananciasAcumuladasPorEfectoDeLaConvergencia; // Añadido
-    private Integer gananciasAcumuladasDiferentesALasGeneradasPorEfectoDeLaConvergencia; // Añadido
+    private Long patrimonio; // Añadido
+    private Long capitalPagado; // Añadido
+    private Long primaDeEmision; // Añadido
+    private Long readquisicionDeInstrumentosDePatrimonioPropio; // Añadido
+    private Long inversionSumplementariaAlCapitalAsignado; // Añadido
+    private Long otrasParticipacionesEnElPatrimonio; // Añadido
+    private Long superavitPorRevaluacion; // Añadido
+    private Long reservaLegal; // Añadido
+    private Long otrasReservas; // Añadido
+    private Long utilidadYorPerdidaDelEjercicio; // Añadido
+    private Long utilidadYorPerdidaAcumulada; // Añadido
+    private Long gananciasAcumuladasPorEfectoDeLaConvergencia; // Añadido
+    private Long gananciasAcumuladasDiferentesALasGeneradasPorEfectoDeLaConvergencia; // Añadido
     private String validacionPatrimonio; // Añadido
     private String validacionGananciaPerdidaNeta; // Añadido
-    private Integer totalPasivoPatrimonio;
+    private Long totalPasivoPatrimonio;
     private String validacionEcuacionPatrimonial;
     private Integer nit;
     private Integer annio;
